Implement reverse assignemnt
+ Implement range_search_assign for assigning elements using range 
search on lsh
+ Fix wall warnings
+ Change printRepresentative output
+ Implement getIndex function in database the returns index of item *
+ Change range_search in hashtable to c instead of previous
+ Add tests for reverse assignment
+ implement closest repr that returns pair of distance and index to/of 
closest representative to item

+++ src/Curve.cpp
@@ -76,7 +76,7 @@ double Curve::distance(Item *p) {
 }
 Item * Curve::clone(void) {
     Curve *p = new Curve(this->getId());
-    for (size_t i = 0; i < this->getSize(); i++) {
+    for (int i = 0; i < this->getSize(); i++) {
         Point * m = this->getPoint(i);
         p->addPoint(m->x,m->y);
     }
